Remove D207A which was causing less going out to C_YUBFEA
I did a study comparison between the CL 3.1.0 studies in the
corroboration folder. One was in the root folder while one was in
another subfolder. I found that they differed in that the one which
showed the low difference comparison was the one without D207A.
The reasoning is that arc is a non project diversion. But then again in
AD_YUBFEA, D207A is ommitted. May need to check later how the results
change when I add D207A in CS2CL